backend/requirements.txt
praw
requests
python-dotenv
boto3
googlemaps
huggingface_hub
google-generativeai
feedparser
